+++ Projectile.h
@@ -1,7 +1,6 @@
 #pragma once
 #include "Movable.h"
 
-
 #define M_PI 3.14159265359f
 
 class Game;
@@ -32,7 +31,7 @@ private:
 	friend class Minion;
 
 public:
-	Projectile(Game* game0, Minion*& target0, int type0, int damage0, int armorPenetration0, int velocity0, Vector2f position);
+	Projectile(Game* game0, Minion*& target0, int type0, int damage0, int armorPenetration0, float velocity0, Vector2f position);
 
 	void move();
